### 2024-09-16 #7238 @editorialbot
[PRE REVIEW]: target-methylseq-qc: a lightweight pipeline for collecting metrics from targeted sequence mapping files.
**Submitting author:** @abhi18av (<a href="http://orcid.org/0000-0002-6402-6993">Abhinav Sharma</a>)
**Repository:** https://github.com/wal-yan/target-methylseq-qc
**Branch with paper.md** (empty if default branch): master
**Version:** v2.0.0
