### 2024-09-18 8cdda70
update docs with the image assets [ci skip]
