+++ docs/output.md
@@ -30,13 +30,6 @@ The pipeline is built using [Nextflow](https://www.nextflow.io/) and processes d
 
 [FastQC](http://www.bioinformatics.babraham.ac.uk/projects/fastqc/) gives general quality metrics about your sequenced reads. It provides information about the quality score distribution across your reads, per base sequence content (%A/T/G/C), adapter contamination and overrepresented sequences. For further reading and documentation see the [FastQC help pages](http://www.bioinformatics.babraham.ac.uk/projects/fastqc/Help/).
 
-![MultiQC - FastQC sequence counts plot](images/mqc_fastqc_counts.png)
-
-![MultiQC - FastQC mean quality scores plot](images/mqc_fastqc_quality.png)
-
-![MultiQC - FastQC adapter content plot](images/mqc_fastqc_adapter.png)
-
-> **NB:** The FastQC plots displayed in the MultiQC report shows _untrimmed_ reads. They may contain adapter sequence and potentially regions with low quality.
 
 ### MultiQC
 
@@ -51,6 +44,10 @@ The pipeline is built using [Nextflow](https://www.nextflow.io/) and processes d
 </details>
 
 [MultiQC](http://multiqc.info) is a visualization tool that generates a single HTML report summarising all samples in your project. Most of the pipeline QC results are visualised in the report and further statistics are available in the report data directory.
+
+
+![MultiQC report generated for target-methylseq-qc, in `picard-profiler` highlighting the refine metrics from targeted sequencing at 10X, 20X, 30X and 50X coverage.\label{results-2a}](multiqc.tiff)
+
 
 Results generated by MultiQC collate pipeline QC from supported tools e.g. FastQC. The pipeline has special steps which also allow the software versions to be reported in the MultiQC output for future traceability. For more information about how to use MultiQC reports, see <http://multiqc.info>.
 
@@ -93,6 +90,12 @@ The `CollectMultipleMetrics` tool allows for the running of multiple metric coll
 * `CollectBaseDistributionByCycle` charts the nucleotide distribution per cycle in a BAM file, producing a chart to assess any systematic errors at specific read positions. Able to be inferred are miscalled bases through reflections in distribution changes, sequencing errors, and trimming quality, find here [Interpretation notes](https://broadinstitute.github.io/picard/command-line-overview.html#CollectBaseDistributionByCycle)
 * `CollectInsertSizeMetrics` provides metrics for validation of library construction, including the distribution of insert size and read orientation of pair ended libraries. Depending on the type of library preparation used these metrics will vary. Output inclues percentages of read pairs in each orientation as a histogram, as well as the insert size distribution as a histogram and txt file.
 * `CollectQualityYieldMetrics` provides metrics about reads passing quality thresholds and Illumina-specific filters. The overall quality of the reads within a BAM file are analysed, and output indicates the total number of bases within reads that pass the minimum base quality score and Illumina quality filter thresholds. Included here are the quality by cycle, quality distribution, and read length metric files.
+
+
+### Bed Filter
+
+
+![Filtered bedGraph file generated using the `bed_filter` mode of target-methylseq-qc.\label{results-2b}](bed_filter.tiff){width=70%}
 
 ### Pipeline information
 
