### 2024-09-20 #7238 @Kevin-Mattheus-Moerman
- [ ] I may have missed it, but can you confirm this project features (automated) testing? If so it may be good to link to this in the README.
- [ ] Could you help me understand the above code report, and potentially add to it in terms of nextflow contributions? What aspects of the report would you say is your core achievement/new contribution? In addition, can you help estimate the lines of code number for the nextflow work? We ask this as some tools exist e.g. to automatically generate JavaScript GUI related code for instance. So any help to judge the "weight/size" of this submission would be appreciated.
