### 2024-09-23 ef2c826
update readme and trigger tests
