### 2024-09-23 910c8ee
update bibliography
