+++ paper/paper.bib
@@ -37,6 +37,7 @@ https://www.nature.com/articles/nbt.3820.pdf},
    month = {2022/05//},
    abstract = {Repository to host tool-specific module files for the Nextflow DSL2 community!},
    url = {https://github.com/nf-core/tools},
+   doi = {10.1038/s41587-020-0439-x},
    year = {2022},
    type = {Computer Program}
 }
