### 2024-09-23 c770855
update readme, bib and docs as per JOSS review comments
