+++ paper/paper.bib
@@ -366,6 +366,6 @@ Proteomics/methods
   month        = aug,
   year         = 2024,
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.13601364},
-  url          = {https://doi.org/10.5281/zenodo.13601364}
+  doi          = {10.5281/zenodo.13597863},
+  url          = {https://doi.org/10.5281/10.5281/zenodo.13597863}
 }
