+++ docs/usage.md
@@ -61,8 +61,10 @@ Whereas the `bed_filter` mode requires a different set of columns in the input s
 
 ### Test profiles
 
-Two built-in test profiles are available in target-methylseq-qc pipeline for each mode of execution. These profiles can be used to run tests on the relevant infrastructure using the bundled test datasets [@test_dataset], helping users to identify and resolve any infrastructural issue before the analysis stage.
+Two built-in test profiles are available in target-methylseq-qc pipeline for each mode of execution. These profiles can be used to run tests on the relevant infrastructure using the bundled test datasets ([published on Zenodo](https://doi.org/10.5281/10.5281/zenodo.13597863)), to help users identify and resolve any infrastructural issue before the analysis stage.
 
+
+**NOTE**: The snippets below assumes you have `docker` on the sever/machine you wish to test the pipeline. For other institutional configs please refer [nf-core/configs](https://nf-co.re/docs/usage/configuration#max-resources) project, which are all applicable to this pipeline.
 
 ```bash
 
