### 2024-09-23 6f2021f
update cff file and bib for proper zenodo doi
