+++ paper/paper.bib
@@ -356,8 +356,8 @@ Proteomics/methods
   year         = 2024,
   publisher    = {Zenodo},
   version      = {v2.0.0},
-  doi          = {10.5281/zenodo.13147688},
-  url          = {https://doi.org/10.5281/zenodo.13147688}
+  doi          = {10.5281/zenodo.8251379},
+  url          = {https://doi.org/10.5281/zenodo.8251379}
 }
 
 @dataset{test_dataset,
