+++ CITATION.cff
@@ -18,7 +18,7 @@ authors:
   given-names: "Patricia"
   orcid: "https://orcid.org/0000-0002-3703-4111"
 title: "wal-yan/target-methylseq-qc"
-version: 1.0.0
-#doi: 10.5281/zenodo.1234
-date-released: 2023-08-10
+version: 2.0.0
+doi: 10.5281/zenodo.8251379
+date-released: 2024-08-01
 url: "https://github.com/wal-yan/target-methylseq-qc"
