### 2024-09-23 #7238 @abhi18av
> @abhi18av Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

Dear @Kevin-Mattheus-Moerman and @csoneson , thank you for your time to evaluate this manuscript!

I have addressed the comments inline.

> * ✅   Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.

Sure, I have updated the DOI for a few more citations, however as some entries don't have an associated publication (such as `picard`) or there's no consensus on how to cite (such as `HTS specification` https://github.com/samtools/hts-specs/issues/179 ), I have simply used the `@online` bib resource annotation for those.

If there's a better way or a JOSS convention to address this, please let us know and we will be happy to accommodate.

> * ✅  I may have missed it, but can you confirm this project features (automated) testing? If so it may be good to link to this in the README.

Ah yes, there are a bunch of [Github actions in the repo](https://github.com/wal-yan/target-methylseq-qc/tree/master/.github/workflows) which are triggered upon relevant events.

In addition, I have added an explanation in the REAMDE for the bundled test dataset which we provide to users for quick testing https://github.com/wal-yan/target-methylseq-qc?tab=readme-ov-file#testing .

> * ✅  Could you help me understand the above code report, and potentially add to it in terms of nextflow contributions? What aspects of the report would you say is your core achievement/new contribution? In addition, can you help estimate the lines of code number for the nextflow work? We ask this as some tools exist e.g. to automatically generate JavaScript GUI related code for instance. So any help to judge the "weight/size" of this submission would be appreciated.

In terms of the `cloc` report from https://github.com/openjournals/joss-reviews/issues/7238#issuecomment-2352844762, I must say that numbers hide the overall big picture, but thank you for raising this.

The principle changes regarding the **implementation logic** are of course in the **Nextflow/Groovy layer**, however as Nextflow is just the DSL for the orchestration of tasks, we have worked on other layers/languages as well.

The **samplesheet check** (written in **Python**) is specific to this pipeline and checks for the overall validity of the samplesheet as a `pre-flight` check, in addition to the **test samplesheet files in CSV format** (`assets/test_samplesheet_bed_filter.csv` and `assets/test_samplesheet_picard_profiler.csv`.

Furthermore, once the analysis is done, the generated results are merged and pushed to `MultiQC` which relies on a customized YAML file (**`assets/multiqc_config.yml`**) in order to present the principal summary report.

Finally in terms of the UI for Nextflow Schema renderers, the **JSON format** has been customized to reflect the principal parameters of the pipeline corresponding to different modes.

I must also highlight that the creation of `test_*` profiles is done in **`conf/*config`** scripts which are also `Groovy/Nextflow` scripts but __do not__ get picked up by `cloc` as any language in the overall counts.

Therefore, kindly take this into consideration 🙏
