### 2024-09-23 #7238 @Kevin-Mattheus-Moerman
@abhi18av thanks for providing those additional details. I have just flagged this submission for a scope review by our editorial board. This is because I need some help to determine if this work is in scope, and if the pipeline/workflow you present meets our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.

The scope review should take about 2 weeks to complete.
