### 2024-09-24 #7238 @abhi18av
Hi @Kevin-Mattheus-Moerman ,

I was trying to understand `cloc` better and I've edited my response here https://github.com/openjournals/joss-reviews/issues/7238#issuecomment-2367633717 to note that `cloc` ( or `tokei` etc) do not take Nextflow files into account therefore the principal implementation logic (in `.nf` and `.config` files) is not really included in the line count reports unfortunately 🤷
