### 2024-09-24 #7238 @Kevin-Mattheus-Moerman
@abhi18av yes I'm aware cloc doesn't count nextflow code lines. Hence I asked you to elaborate. If you can help count/estimate these lines yourself that would be helpful.
