### 2024-11-19 #7238 @csoneson
@abhi18av In the meanwhile, I also noticed that your paper is a bit on the long side for JOSS standards (recommended length is 250-1000 words, see [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)). I think you could, for example, move some of the parts related to the installation and available parameters to the GitHub repository (again, see the link above for more details about what we expect a JOSS paper to contain).

Also, could you please add a country to affiliation (1). Thanks!
