### 2024-11-21 aa22bf5
update the affiliations and shorten the prose
