+++ paper/paper.bib
@@ -338,12 +338,6 @@ Proteomics/methods
 }
 
 
-@online{targetmethylseqqc_website,
-    title = {target-methylseq-qc website},
-    year = {2024},
-    url = {https://wal-yan.github.io/target-methylseq-qc/usage.html},
-}
-
 
 @software{targetmethylseqqc,
   author       = {Sharma, Abhinav and
@@ -368,4 +362,12 @@ Proteomics/methods
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.13597863},
   url          = {https://doi.org/10.5281/10.5281/zenodo.13597863}
+}
+
+
+@online{targetmethylseqqc_website,
+    title = {target-methylseq-qc website},
+    year = {2024},
+    doi  = {10.5281/zenodo.8251379},
+    url = {https://wal-yan.github.io/target-methylseq-qc/usage.html},
 }
