### 2024-11-21 #7238 @abhi18av
Hi @Kevin-Mattheus-Moerman  and @csoneson ,

We are glad to hear the news that the our submission is progressing in the review pipeline.

@csoneson  I have accommodated your suggestions and trimmed the sections regarding download, test profiles etc from the content, however I think that the content of tables is what increases the word count overall. I'm not sure if there's a way to instruct the editorialbot not to count the table as plain text 🤔

In any case, we would be happy to accommodate any further changes you feel fit for the publication.
