### 2024-11-28 #7238 @abhi18av
> ⚠️ Note to editors: If these papers look like they might be a good match, click through to the review issue for that paper and invite one or more of the authors before considering asking the reviewers of these papers to review again for JOSS. (From https://github.com/openjournals/joss-reviews/issues/7238#issuecomment-2496240213)

Hi @csoneson, just to  confirm something regarding the above message, do we need to propose some reviewers ourselves or this is already in progress?
