### 2024-12-19 #7238 @csoneson
👋🏻  @daissi, @kdm9, @johandahlberg - would one of you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> target-methylseq-qc: a lightweight pipeline for collecting metrics from targeted sequence mapping files
https://github.com/openjournals/joss-reviews/issues/7238

More information about the review process can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Thanks in advance!
