### 2025-01-29 fb98884
update documentation and fix images
