### 2025-02-02 wal-yan/picard-profiler#5 @abhi18av
Thank you @telatin, for a thorough review. Really appreciate your time 🙏
