### 2025-02-14 wal-yan/target-methylseq-qc#9 @sridhar0605
JOSS-Statement of need
[Statement of need]
I see that the original [methylseq](https://nf-co.re/methylseq/2.6.0/docs/output/#qualimap) has a baked in qualimap QC tool, could you please explain what additional QC is `target-methylseq-qc` offering that is going to helpful.
