### 2025-03-12 wal-yan/picard-profiler#9 @abhi18av
Hi @sridhar0605 , thank you for this question.

I've just come back from a break and still catching up; please allow me some time to circle back on this question after discussing with co-authors.

CC @agudeloromero
