### 2025-03-19 wal-yan/picard-profiler#9 @abhi18av
Hi @sridhar0605

Apologies for the late response here, but happy to respond back with our answer after some internal discussions.

> I see that the original [methylseq](https://nf-co.re/methylseq/2.6.0/docs/output/#qualimap) has a baked in qualimap QC tool,

Indeed, this is correct. The point we wish to highlight is that methylseq pipeline includes the alignment step and therefore the Qualimap QC is used for evaluating the quality of generated BAM files.

However, in the case of the `target-methylseq-qc` pipeline, (i) the `picard_profiler` mode uses the pre-existing BAM files to calculate the coverage with respect to an existing reference FASTA (ii) and the `bed_filter` mode relies directly on the raw BED file to generate the BED file for for the provided sequencing panel which in our study was [Human TWIST methylseq panel](https://www.twistbioscience.com/products/ngs/fixed-panels/human-methylome-panel).

Therefore, with the `target-methylseq-qc` pipeline, we aimed to diversify the post-processing steps after the initial `methylseq` pipeline for targeted sequencing.
