### 2025-03-24 #7608 @csoneson
Thanks @sridhar0605.
@telatin - just to confirm, do I understand correctly from your comment in https://github.com/wal-yan/target-methylseq-qc/issues/5 that you are also happy with the submission, or do you have outstanding concerns?
@abhi18av - I will also have a quick look through the submission, will get back to you shortly!
