### 2025-04-02 1a98c4b
Suggested edits for paper and bibliography
