+++ paper/paper.md
@@ -56,7 +56,7 @@ Next-generation targeted genome sequencing allows the analysis of regions of int
 
 The `target-methylseq-qc` pipeline, when used in the `picard_profiler` mode, accepts inputs in various alignment formats, including SAM, BAM and CRAM files [@hts_spec]. Additionally, to refine the metrics to the target regions the inclusion of a FASTA reference file and BED intervals file is required. Upon completion of the analysis, a MultiQC report [@ewels_multiqc_2016] will be generated, encompassing the updated sequencing coverage data for the targeted regions with some extras. The `picard_profiler` mode of the pipeline integrates Picard metrics from GATK Picard tools [@mckenna_genome_2010; @Picard2019toolkit], using two specific metrics: (i) collecthsmetrics [@picard_collecthsmetrics_2019], which relies upon the hybrid-selection technique to capture exon sequences for targeted sequencing experiments; and (ii) collectmultiplemetrics [@picard_collectmultiplemetrics_2021], which captures closely related metrics such as alignment summary, insert size, and quality score.
 
-The `bed_filter` mode of the pipeline uses the bedtools [@quinlan_2010] filter command to filter the `bedGraph` files output by the nf-core/methylseq workflow [@methylseq] using a reference bed panel, in this case the Twist Human Methylome panel (https://www.twistbioscience.com/resources/data-files/twist-human-methylome-panel-target-bed-file) and best practices [@twist_methylome, @twist_methylome_technote]. Filtering raw BED files to the targeted regions is crucial because it ensures that the analysis focuses on specific genomic targets accurately and efficiently. This step minimizes the inclusion of off-target sequences and reduces the potential for including sequencing artifacts, which can be introduced during capture-based targeted sequencing processes. Downstream analyses from the filtered BED files will enable the calculation of CpG ratios and the testing for differentially methylated cytosines (DMCs) or regions (DMRs).
+The `bed_filter` mode of the pipeline uses the bedtools [@quinlan_2010] filter command to filter the `bedGraph` files output by the nf-core/methylseq workflow [@methylseq] using a reference bed panel, in this case the Twist Human Methylome panel (https://www.twistbioscience.com/resources/data-files/twist-human-methylome-panel-target-bed-file) and best practices [@twist_methylome;@twist_methylome_technote]. Filtering raw BED files to the targeted regions is crucial because it ensures that the analysis focuses on specific genomic targets accurately and efficiently. This step minimizes the inclusion of off-target sequences and reduces the potential for including sequencing artifacts, which can be introduced during capture-based targeted sequencing processes. Downstream analyses from the filtered BED files will enable the calculation of CpG ratios and the testing for differentially methylated cytosines (DMCs) or regions (DMRs).
 
 Regardless of the usage mode of the pipeline, the final MultiQC report automatically collates the relevant reports from FastQC [@andrews_fastqc_2010], bedtools and Picard tools in an HTML document, which could be shared with collaborators or added as supplementary material in publications.
 
