### 2025-04-02 #7608 @csoneson
👋🏻 @abhi18av - I have looked through the submission and made a [PR](https://github.com/wal-yan/target-methylseq-qc/pull/10) with suggestions for edits to the paper and bibliography. Please merge if you agree, and generate a new proof with `@editorialbot generate pdf`. In addition to these edits, a few things were a bit unclear to me when reading the paper:

* line 30: "with some extras" - what does this refer to?
* line 39: "in this case the Twist Human Methylome panel" - what does "in this case" refer to here? Is the pipeline restricted to human samples, or is this merely an example?
* line 90-91: it was not clear to me what it meant that you "enabled the metrics collection for 10x, 20x, 30x and 50x coverage"
* I would recommend removing the `Pipeline setup` section (especially since your paper is a bit above the length recommended in the JOSS guidelines). This information fits better in the GitHub repository/documentation (where it can be kept up to date).
