### 2025-04-04 #7608 @abhi18av
Thank you for the review @csoneson - we're happy to make the suggested changes and merge this PR :)

However, we'd also like to request a bit of an extension since internally there's an ongoing conversation regarding one particular statement regarding the study context and as soon as we have a decision, I'd batch all changes together and rebuild the manuscript.

CC @agudeloromero
