### 2025-04-16 3a4fb14
Suggested edits for paper and bibliography (JOSS review) (#10)
* Suggested edits for paper and bibliography

* A few more edits

* Fix citation

* Consistent case for x in coverage

---------

Co-authored-by: Abhinav Sharma <abhi18av@users.noreply.github.com>
