+++ paper/paper.bib
@@ -127,32 +127,32 @@ Proteomics/methods
 
 
 @software{methylseq,
-  author       = {Phil Ewels and
-                  Patrick Hüther and
-                  Edmund Miller and
-                  Sateesh\_Peri and
-                  Nathan Spix and
-                  nf-core bot and
-                  Alexander Peltzer and
-                  Sven F. and
-                  Johannes Alneberg and
-                  Maxime U Garcia and
-                  Felix Krueger and
-                  Paolo Di Tommaso and
-                  Matthias Hörtenhuber and
-                  Vikesh Ajith and
-                  Colin Davenport and
-                  Harshil Patel and
-                  Wassim Salam and
-                  Noe Cochetel and
+  author       = {Ewels, Philip A and
+                  Hüther, Patrick and
+                  Miller, Edmund and
+                  Peri, Sateesh and
+                  Spix, Nathan and
+                  {nf-core bot} and
+                  Peltzer, Alexander and
+                  {Sven F.} and
+                  Alneberg, Johannes and
+                  Garcia, Maxime U and
+                  Krueger, Felix and
+                  {Di Tommaso}, Paolo and
+                  Hörtenhuber, Matthias and
+                  Ajith, Vikesh and
+                  Davenport, Colin and
+                  Patel, Harshil and
+                  Salam, Wassim and
+                  Cochetel, Noe and
                   Alessia and
-                  Guillaume Devailly and
-                  Robert Syme and
-                  Simon Pearce and
+                  Devailly, Guillaume and
+                  Syme, Robert and
+                  Pearce, Simon and
                   VIJAY and
-                  Kevin Menden and
+                  Menden, Kevin and
                   ltosti-tagomics and
-                  Noirot Céline},
+                  Noirot, Céline},
   title        = {nf-core/methylseq: Huggy mollusc},
   month        = jan,
   year         = 2024,
@@ -283,7 +283,7 @@ Proteomics/methods
     pages = {3047--3048},
     number = {19},
     journaltitle = {Bioinformatics},
-    author = {Ewels, Philip and Magnusson, Måns and Lundin, Sverker and Käller, Max},
+    author = {Ewels, Philip A and Magnusson, Måns and Lundin, Sverker and Käller, Max},
     urldate = {2023-10-05},
     date = {2016-10-01},
     langid = {english},
@@ -300,7 +300,7 @@ Proteomics/methods
     number = {3},
     journaltitle = {Nature Biotechnology},
     shortjournal = {Nat Biotechnol},
-    author = {Ewels, Philip A. and Peltzer, Alexander and Fillinger, Sven and Patel, Harshil and Alneberg, Johannes and Wilm, Andreas and Garcia, Maxime Ulysse and Di Tommaso, Paolo and Nahnsen, Sven},
+    author = {Ewels, Philip A and Peltzer, Alexander and Fillinger, Sven and Patel, Harshil and Alneberg, Johannes and Wilm, Andreas and Garcia, Maxime Ulysse and Di Tommaso, Paolo and Nahnsen, Sven},
     urldate = {2023-10-05},
     date = {2020-03},
     langid = {english},
