### 2025-04-25 #7608 @csoneson
Thanks @abhi18av for updating the paper, and sorry for the delayed response. I have one more comment about the following modified sentences:

> Although target-methylseq-qc was originally created for calculating sequencing coverage in target sequencing as a follow-up step to the `nf-core/methylseq` pipeline [@methylseq], aimed at identification of children at the risk of developing asthma, prior to its onset [@patricia2023-aerial]. The pipeline's versatility allows for extending its application to other sequencing panels from various next-generation methods.

Could you reformulate this either by merging it back into one sentence (otherwise the "Although..." part is left dangling in the first sentence), or by removing the "Although" and perhaps adding a "However, " in the second sentence.
