### 2025-04-25 8c6870e
accommodate prose suggestion from the editor
