### 2025-04-25 #7608 @abhi18av
Hi @csoneson , we have accommodate this feedback here https://github.com/wal-yan/target-methylseq-qc/commit/8c6870e0f58aff529ca1ef679f3fcdac89e9a16b

Please let us know if there's anything else that can be improved and we'd happy to accommodate that.

Have a nice weekend ahead!
