### 2025-04-27 #7608 @csoneson
Great, thanks @abhi18av - I have made another small [PR](https://github.com/wal-yan/target-methylseq-qc/pull/11) to fix a citation that didn't format properly. In addition, I just noticed that the `@patricia2023-aerial` bib item doesn't seem to exist, could you please add the proper reference here?

Once those things are fixed I think we are good to move on to the next steps. I will make another post below outlining those.
