+++ paper/paper.bib
@@ -392,3 +392,17 @@ Cited References Count:25},
    year = {2022},
    type = {Journal Article}
 }
+
+@inproceedings{patricia2023-aerial,
+   author = {Agudelo-Romero, Patricia and Iosifidis, Thomas and Kicic-Starcevich, Elizabeth and Hancock, David and Caparros-Martin, Jose and Martino, David and Stick, Stephen},
+   title = {Nasal and Amnion Methylomes: Biomarkers for Smoke Exposure and Maternal Asthma},
+   booktitle = {Journal of Allergy and Clinical Immunology},
+   publisher = {Elsevier},
+   volume = {153},
+   pages = {AB243},
+   note = {doi: 10.1016/j.jaci.2023.11.779},
+   ISBN = {0091-6749},
+   DOI = {10.1016/j.jaci.2023.11.779},
+   url = {https://doi.org/10.1016/j.jaci.2023.11.779},
+   type = {Conference Proceedings}
+}
