+++ paper/paper.bib
@@ -404,5 +404,6 @@ Cited References Count:25},
    ISBN = {0091-6749},
    DOI = {10.1016/j.jaci.2023.11.779},
    url = {https://doi.org/10.1016/j.jaci.2023.11.779},
+   year = {2024},
    type = {Conference Proceedings}
 }
