### 2025-04-29 #7608 @abhi18av
Hi @csoneson ,

Here's our follow up for the next steps in https://github.com/openjournals/joss-reviews/issues/7608#issuecomment-2833357071

[Tag `v2.1.0`](https://github.com/wal-yan/target-methylseq-qc/releases/tag/v2.1.0) reflects all the changes made since the review was initiated.

- [x] Archive the release on Zenodo/figshare/etc and post the DOI here.`v2.1.0` is

https://doi.org/10.5281/zenodo.15302404

License on Github for `v2.1.0` (MIT License)

https://github.com/wal-yan/target-methylseq-qc/blob/fb2b4f4059ec6a759407ac68038d129eb2c0bda4/LICENSE#L1

License on Zenodo for `v2.1.0`
