### 2025-04-30 #7608 @abhi18av
Hi @csoneson , we've updated the name of the zenodo record as you recommended.
