### 2025-05-01 #7608 @csoneson
@abhi18av Ok, this looks all good to me. Just for completeness, at this point I will conclude based on @telatin ticking all boxes in the checklist and noting 'All good for me' in https://github.com/wal-yan/target-methylseq-qc/issues/5#issuecomment-2628961715 that there are indeed no outstanding issues on their side.
I will now hand over to the track AEiC for the last steps. Thanks for submitting to JOSS!
