### 2025-05-09 #7608 @csoneson
@abhi18av - could you try to change `ISBN` to `ISSN` [here](https://github.com/wal-yan/target-methylseq-qc/blob/2810396a1a0c44f114de086f411e30818f3f95df/paper/paper.bib#L404), to see if that solves the issue above?
