+++ paper/paper.bib
@@ -401,7 +401,7 @@ Cited References Count:25},
    volume = {153},
    pages = {AB243},
    note = {doi: 10.1016/j.jaci.2023.11.779},
-   ISBN = {0091-6749},
+   ISSN = {0091-6749},
    DOI = {10.1016/j.jaci.2023.11.779},
    url = {https://doi.org/10.1016/j.jaci.2023.11.779},
    year = {2024},
