### 2025-05-09 #7608 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
