### 2025-05-17 #7608 @Kevin-Mattheus-Moerman
👉  The paper features both British (e.g. `behaviour`, `standardised`, `summarised`) and American English (`minimizes`, `utilized`, `standardized`, `standardizing`, `initialization`, `summarize`) please edit so one is used consistently.
