### 2025-05-19 e925880
docs: fix typos in paper.md
Corrected spelling errors in the documentation, including "standardised" to "standardized," "behaviour" to "behavior," and "summarised" to "summarized." These changes improve consistency and adhere to American English conventions.
