### 2025-05-19 #7608 @abhi18av
Hi @Kevin-Mattheus-Moerman, I have addressed the typos and switched entirely to American English.

https://github.com/wal-yan/target-methylseq-qc/commit/e925880436e7a155f68551d99d15823392157fac
